pyproject.toml
[project]
name = "cw-habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-extensions (>=4.1,<5.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "python-telegram-bot (>=22.4,<23.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
mypy = "^1.18.1"


[tool.poetry.group.dev.dependencies]
pillow = "^11.3.0"
dotenv = "^0.9.9"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
autopep8 = "^2.3.2"
ipython = "^9.5.0"
phonenumbers = "^9.0.14"
django-phonenumber-field = "^8.1.0"
redis = "^6.4.0"
django-redis = "^6.0.0"
django = "^5.2.6"
django-select2 = "^8.4.1"
djangorestframework = "^3.16.1"
requests = "^2.32.5"
celery = "^5.5.3"
django-celery-beat = "^2.8.1"
django-celery-results = "^2.6.0"

[tool.poetry.dependencies]
python = ">=3.13,<4"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
multi_line_output = 3
